Script to translate the model file to PMML
file as input and output a directory with pmml files. Each file can be given as input to
pmml consumers. You can test the files using jpmml-evaluator. Please refer to this the
data mining group's website for the pmml specifcation (http://dmg.org/pmml/pmml-v4-3.html).
Also refer to the README.md file in the pmml directory to see how to test the pmml output
with jpmml-evaluator.