[CLI][python-package][docs] Add n_estimators as num_iteration alias (#1079)
* Add n_estimators as num_iteration alias

Scikit-Learn's ensemble methods use the term `n_estimators` for the number of
iterations of training models. To make it more accessible for newcomers who are
familiar with Scikit-Learn, it would help if the Parameters page mentioned
`n_estimators` and what parameter that maps to within LightGBM's API.

Addresses discussion brought up in #954

* Add n_estimators as num_iterations alias

Adds n_estimators as an alias for num_iterations in the CLI as well as Python
libs. Additionally bumps the default for n_estimators in the Sklearn API to 100
estimators.