Automatically generate config_auto.cpp from config.h at the build time
Because src/io/config_auto.cpp is automatically generated from config.h it might
be generated at the build stage, and there is no need to launch
helper/parameter_generator.py manually.

It adds Python as a build dependency, but it is already impossible to
build the project without python if you changed anything in the config.h
file.