Update GPU-Windows.rst, missing bootstrap.bat gcc
Added line 190, "bootstrap.bat gcc"
This step is necessary in order to complete the following steps of "b2 install...". b2 will not be present in C:\boost\boost_1_63_0 unless "bootstrap.bat gcc" is run there first (this is in addition to the "bootstrap.bat gcc" in line 187).
I have also added another "bootstrap" to line 241 (and missing parenthesis) to make it match the update.