Iteration logic consistent with python-pacakge
As per discussion on GH-#5051 and GH-#5193 the python package does not stop
training if a single leaf tree (stupm) is found and relies on early
stopping methods to stop training. This commits removes the finish
condition on training based on the result of `TrainOneIter()` and sets
the `is_finished` flag on early stopping alone.