lightgbm_fixes.py
In this commit, there has been some significant changes that have been made to the script in a bid to improve its functionality and to make it as Pythonic as possible. The parameter `X` used in the `_check_sample_weight` function has been renamed to `input_array` to make it compliant with Python naming guidelines. It helps in making the code more readable and Same thing in more of a systematic manner. Also, one more argument which is not required in the function i. e. , `dtype` has been removed from the function. This cleanup remove some noise from the function signature, makes the code more clear and can avoid confusion. 
 
 Moreover, the class named as ‘pd_Series’ has been changed to ‘PdSeries’ to adhere to the naming convention of Python classes where the first letter of each word in the class name has to be capitalised. These updates combined together improve the code’s readability, reduce code complexity and improve the quality of the code in accordance with the best practices.