BUGFIX of non deterministic error for multiple model trainings
before when training multiple models within one bash script some of them failed to finish without useful error messages, but just when rounding the threshold values.
Using the rounded value in the CalculateSplitMemory function solved the issue, as when called from SerialTreeLearner::ComputeBestSplitForFeature a not rounded value was compared to the thresholds_used_global_ vector, only containing rounded values.

Still the actual reason for the bug is not clear as this scenario was actually covered by the previous code